Student, Department, Faculty, Person classes added

+++ src/mum/mpp_lab/lab1/Student.java
@@ -6,7 +6,7 @@ public class Student extends Person {
     private double gpa;
     private ArrayList<Course> courses;
 
-    public Student(String name, String phone,int age, int gpa){
+    public Student(String name, String phone,int age, double gpa){
         this.setName(name);
         this.setPhone(phone);
         this.setAge(age);
